+++ src/main/kotlin/app/revanced/patches/youtube/layout/hideinfocards/fingerprints/InfocardsIncognitoParentFingerprint.kt
@@ -1,10 +1,10 @@
-package app.revanced.patches.youtube.ad.infocards.fingerprints
+package app.revanced.patches.youtube.layout.hideinfocards.fingerprints
 
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
 import app.revanced.patcher.extensions.or
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-import app.revanced.patches.youtube.ad.infocards.annotations.HideInfocardsCompatibility
+import app.revanced.patches.youtube.layout.hideinfocards.annotations.HideInfocardsCompatibility
 import org.jf.dexlib2.AccessFlags
 
 @Name("infocards-incognito-parent-fingerprint")
