+++ src/main/kotlin/app/revanced/patches/youtube/layout/hideinfocards/fingerprints/InfocardsMethodCallFingerprint.kt
@@ -1,10 +1,10 @@
-package app.revanced.patches.youtube.ad.infocards.fingerprints
+package app.revanced.patches.youtube.layout.hideinfocards.fingerprints
 
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
-import app.revanced.patches.youtube.ad.infocards.annotations.HideInfocardsCompatibility
-import app.revanced.patches.youtube.ad.infocards.resource.patch.HideInfocardsResourcePatch
+import app.revanced.patches.youtube.layout.hideinfocards.annotations.HideInfocardsCompatibility
+import app.revanced.patches.youtube.layout.hideinfocards.resource.patch.HideInfocardsResourcePatch
 import org.jf.dexlib2.Opcode
 import org.jf.dexlib2.iface.instruction.WideLiteralInstruction
